Fix the button layout. Extract styles from hook as it leads to error when used without grid provider (columns);
