+++ src/main/topLevelPages/myDashboard/DashboardWithMemberships/DashboardSpaces/DashboardSpaces.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from 'react-i18next';
-import { Paper, Button, Avatar } from '@mui/material';
+import { Paper, Button, Avatar, useTheme, useMediaQuery, Theme } from '@mui/material';
 import { Card } from '@mui/material';
 import { DoubleArrowOutlined } from '@mui/icons-material';
 import Gutters from '@/core/ui/grid/Gutters';
@@ -10,14 +10,15 @@ import RouterLink from '@/core/ui/link/RouterLink';
 import { Caption, Tagline } from '@/core/ui/typography';
 import { MyMembershipsDialog } from '@/main/topLevelPages/myDashboard/myMemberships/MyMembershipsDialog';
 import PageContentBlock from '@/core/ui/content/PageContentBlock';
-import JourneyTile from '@/domain/journey/common/JourneyTile/JourneyTile';
+import JourneyTile, { RECENT_JOURNEY_CARD_ASPECT_RATIO } from '@/domain/journey/common/JourneyTile/JourneyTile';
 import { SpacePrivacyMode, VisualType } from '@/core/apollo/generated/graphql-schema';
 import { defaultVisualUrls } from '@/domain/journey/defaultVisuals/defaultVisualUrls';
 import { useDashboardSpaces } from './useDashboardSpaces';
 import { gutters } from '@/core/ui/grid/utils';
-import { useEffect } from 'react';
+import { useEffect, useMemo } from 'react';
 import { LoadingButton } from '@mui/lab';
 import { Actions } from '@/core/ui/actions/Actions';
+import { useColumns } from '@/core/ui/grid/GridContext';
 
 const DASHBOARD_MEMBERSHIPS_ALL = 100; // hardcoded limit for expensive query
 
@@ -27,18 +28,63 @@ const DashboardSpaces = () => {
     hasMore,
     loading,
     fetchSpaces,
-    cardColumns,
     isDialogOpen,
-    visibleSpaces,
     selectedSpaceIdx,
     selectedSpaceName,
-    styles: { loader, spaceCard, spaceTitle, spaceTagline, spaceCardMedia, exploreAllButton, titleAndDescContainer },
-
     handleDialogOpen,
     handleDialogClose,
   } = useDashboardSpaces();
 
+  const theme = useTheme();
+
+  const styles = {
+    loader: {
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+
+      width: '100%',
+      height: '100%',
+    },
+
+    spaceCard: {
+      borderBottomLeftRadius: 0,
+      borderBottomRightRadius: 0,
+    },
+
+    spaceCardMedia: {
+      height: '180px',
+      minWidth: '100%',
+      objectFit: 'cover',
+    },
+
+    titleAndDescContainer: {
+      borderBottomLeftRadius: 16,
+      borderBottomRightRadius: 16,
+      backgroundColor: theme.palette.primary.main,
+    },
+
+    spaceTitle: {
+      color: theme.palette.background.paper,
+    },
+
+    spaceTagline: {
+      color: theme.palette.background.paper,
+    },
+
+    exploreAllButton: {
+      textTransform: 'none',
+      border: `1px solid ${theme.palette.divider}`,
+      aspectRatio: RECENT_JOURNEY_CARD_ASPECT_RATIO,
+    },
+  };
+
   const { t } = useTranslation();
+
+  const columns = useColumns();
+  const isMobile = useMediaQuery<Theme>(theme => theme.breakpoints.down('sm'));
+  const cardColumns = useMemo(() => (isMobile ? columns / 2 : columns / 4), [isMobile, columns]);
+  const visibleSpaces = Math.max(1, Math.floor(columns / 2) - 1);
 
   useEffect(() => {
     fetchSpaces(); // with default limit
@@ -46,7 +92,7 @@ const DashboardSpaces = () => {
 
   if (loading) {
     return (
-      <PageContentBlock style={loader}>
+      <PageContentBlock style={styles.loader}>
         <Loading />
       </PageContentBlock>
     );
@@ -66,22 +112,22 @@ const DashboardSpaces = () => {
         return (
           <PageContentBlock key={id}>
             <Gutters component={RouterLink} to={profile?.url} disableGap disablePadding>
-              <Card style={spaceCard}>
+              <Card style={styles.spaceCard}>
                 <Avatar
                   variant="square"
-                  sx={spaceCardMedia}
+                  sx={styles.spaceCardMedia}
                   alt={profile?.displayName}
                   src={profile?.spaceBanner?.uri || defaultVisualUrls[VisualType.Banner]}
                 />
               </Card>
 
-              <Gutters gap={gutters(0.3)} padding={gutters(0.3)} style={titleAndDescContainer}>
-                <PageTitle textAlign="center" style={spaceTitle}>
+              <Gutters gap={gutters(0.3)} padding={gutters(0.3)} style={styles.titleAndDescContainer}>
+                <PageTitle textAlign="center" style={styles.spaceTitle}>
                   {profile?.displayName}
                 </PageTitle>
 
                 {tagline && (
-                  <Tagline textAlign="center" fontStyle="italic" color={spaceTagline.color}>
+                  <Tagline textAlign="center" fontStyle="italic" color={styles.spaceTagline.color}>
                     {tagline}
                   </Tagline>
                 )}
@@ -119,7 +165,7 @@ const DashboardSpaces = () => {
                   <GridItem columns={cardColumns}>
                     <Paper
                       component={Button}
-                      sx={exploreAllButton}
+                      sx={styles.exploreAllButton}
                       endIcon={<DoubleArrowOutlined />}
                       onClick={handleDialogOpen(idx, profile?.displayName)}
                     >
@@ -137,8 +183,8 @@ const DashboardSpaces = () => {
         );
       })}
       {hasMore && (
-        <Actions>
-          <LoadingButton loading={loading} onClick={() => fetchSpaces(DASHBOARD_MEMBERSHIPS_ALL)}>
+        <Actions justifyContent="center" sx={{ width: '100%' }}>
+          <LoadingButton variant="contained" loading={loading} onClick={() => fetchSpaces(DASHBOARD_MEMBERSHIPS_ALL)}>
             {t('buttons.load-more')}
           </LoadingButton>
         </Actions>
